requirements.txt
langchain
python-dotenv
ipykernal
langchain-openai
beautifulsoup4
langchain-community
faiss-cpu
streamlit
